Moved the Cypress folder into the same directory as other tests (tests). All template tests were moved under a new directory called template-tests, while the e2e directory was left for e2e tests. Additionally, accessibility tests now have their own directory at the same level as other test directories. Configuration changes were made because changing the directory structure affected the functionality of Cypress Axe
