+++ tests/cypress/accessibility-tests/axe-about.cy.js
@@ -1,11 +1,12 @@
-import { accessibilityTestRunner } from '../../support/accessibility.js'
+import { accessibilityTestRunner } from '../support/accessibility.js'
+import 'cypress-axe';
 
 /* If you want the test to be skipped, add a skip command after the describe part:
     - test enabled: describe('Check accessibility of ...
     - test to be skipped: describe.skip('Check accessibility of ... */
-describe.skip('Check accessibility of the vocab-home page', () => {
+describe('Check accessibility of the about page', () => {
   before(() => {
-    cy.visit('/agrovoc/fi/')
+    cy.visit('/fi/about')
     cy.injectAxe()
   })
