+++ tests/cypress/accessibility-tests/axe-concept.cy.js
@@ -1,6 +1,13 @@
+// When you start creating tests, remove the following comment(s).
+// import { accessibilityTestRunner } from '../support/accessibility.js'
+// import 'cypress-axe';
+
 /* If you want the test to be skipped, add a skip command after the describe part:
     - test enabled: describe('Check accessibility of ...
     - test to be skipped: describe.skip('Check accessibility of ... */
 describe.skip('Check accessibility of the concept page', () => {
   it('no-op test', () => {})
+
+  // When you start creating tests, remove the following comment(s).
+  // accessibilityTestRunner()
 })
