+++ tests/cypress/template-tests/global-search.cy.js
