+++ tests/cypress/template-tests/vocab-home.cy.js
