+++ tests/cypress/template-tests/vocab-search.cy.js
